refactoring 2
