+++ src/Components/Weather/Weather.js
@@ -1,4 +1,4 @@
-import { View, Text, Image } from "react-native";
+import { ActivityIndicator, View, Text, Image } from "react-native";
 import { useState, useEffect } from "react";
 
 import useGetWeather from "../../customHooks/useGetWeather.js";
@@ -7,11 +7,13 @@ import styles from "./Weather.style";
 
 
 const Weather = ({ cityName }) => {
+    console.log("call1")
     const { info, isWeatherLoad } = useGetWeather(cityName);
     const [tip, setTip] = useState("test");
+    console.log("call2")
 
     useEffect(() => {
-        if (info.temp >= 20 && (info.weatherId >= 800 && info.weatherId < 804)) {
+        /*if (info.temp >= 20 && (info.weatherId >= 800 && info.weatherId < 804)) {
             setTip("Ideal weather for outside exercise");
         } else if (info.temp < 15 && (info.weatherId >= 800 && info.weatherId < 804)) {
             setTip("Weather is suitable for outside exercise but we recommend to take good clothes");
@@ -19,9 +21,27 @@ const Weather = ({ cityName }) => {
             setTip("We recommend to go out but be prepared for possible rain");
         } else {
             setTip("Today we recommend to stay inside")
+        }*/
+        switch (true) {
+            case info.temp >= 20 && (info.weatherId >= 800 && info.weatherId < 804):
+                setTip("Ideal weather for outside exercise");
+                break;
+
+            case info.temp < 15 && (info.weatherId >= 800 && info.weatherId < 804):
+                setTip("Weather is suitable for outside exercise but we recommend to take good clothes");
+                break;
+
+            case info.temp >= 15 && info.weatherId == 804:
+                setTip("We recommend to go out but be prepared for possible rain");
+                break;
+
+            default:
+                setTip("Today we recommend to stay inside");
+
+
         }
     }, [isWeatherLoad])
-
+    console.log("call3")
     return (
         <View style={styles.rootContainer}>
             {isWeatherLoad ?
@@ -33,7 +53,7 @@ const Weather = ({ cityName }) => {
                     <Text style={styles.tipContainer}><Text style={styles.ourTip}>Our tip: </Text>{tip}</Text>
                 </>
                 :
-                <Text>Loading...</Text>
+                <ActivityIndicator />
             }
         </View>
     )
